backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
